+++ src/main.ts
@@ -8,7 +8,7 @@ import { createApp } from 'vue';
 import App from './App.vue';
 import router from './services/router';
 import './style.css';
-import pinia from './services/store';
+import { createPinia } from 'pinia';
 
 hljs.registerLanguage('json', json);
 
@@ -21,9 +21,10 @@ hljs.registerLanguage('json', json);
 	},
 };
 
+const pinia = createPinia();
 const app = createApp(App);
 
-app.use(router);
 app.use(pinia);
+app.use(router);
 app.use(hljsVuePlugin);
 app.mount('#app');